engine/requirements.txt
pandas
streamlit
python-dotenv
pdfminer.six
deepagents
langchain-openai
langchain
langchain-community
langchain-core
python-dotenv
faiss-cpu
fastapi
uvicorn
docling